test(autodiff): log_softmax backward 参照実装の減算を f64 で統一する
テスト内に複製した参照式が積を f32 に丸めてから減算していたため、
grad::log_softmax_vjp_along・CUDA カーネルが減算まで f64 で保持する
方式と不一致だった。減算まで f64 のまま行い最後に 1 回だけ f32 へ
downcast する形へ揃える（同型の指摘が crates/backend-cuda 側にも
あったため併せて修正）。

+++ crates/autodiff/tests/log_softmax_backward_dispatch.rs
@@ -243,7 +243,7 @@ fn reference_log_softmax_backward(
             }
             for c in 0..axis_len {
                 let idx = (o * axis_len + c) * inner + i;
-                dx[idx] = g[idx] - (y[idx].exp() as f64 * sum) as f32;
+                dx[idx] = (g[idx] as f64 - y[idx].exp() as f64 * sum) as f32;
             }
         }
     }
